+++ codec-classes-quic/src/main/java/io/netty/incubator/codec/quic/HmacSignQuicConnectionIdGenerator.java
@@ -16,16 +16,8 @@
 package io.netty.incubator.codec.quic;
 
 import java.nio.ByteBuffer;
-import java.security.InvalidKeyException;
-import java.security.NoSuchAlgorithmException;
-import java.security.SecureRandom;
-import java.util.Arrays;
 
-import io.netty.util.concurrent.FastThreadLocal;
 import io.netty.util.internal.ObjectUtil;
-
-import javax.crypto.Mac;
-import javax.crypto.spec.SecretKeySpec;
 
 /**
  * A {@link QuicConnectionIdGenerator} which creates new connection id by signing the given input
@@ -33,20 +25,6 @@ import javax.crypto.spec.SecretKeySpec;
  */
 final class HmacSignQuicConnectionIdGenerator implements QuicConnectionIdGenerator {
     static final QuicConnectionIdGenerator INSTANCE = new HmacSignQuicConnectionIdGenerator();
-
-    private static final FastThreadLocal<Mac> MACS = new FastThreadLocal<Mac>() {
-        @Override
-        protected Mac initialValue() {
-            return newMac();
-        }
-    };
-
-    private static final String ALGORITM = "HmacSHA256";
-    private static final byte[] randomKey = new byte[16];
-
-    static {
-        new SecureRandom().nextBytes(randomKey);
-    }
 
     private HmacSignQuicConnectionIdGenerator() {
     }
@@ -57,31 +35,13 @@ final class HmacSignQuicConnectionIdGenerator implements QuicConnectionIdGenerat
                 "HmacSignQuicConnectionIdGenerator should always have an input to sign with");
     }
 
-    private static Mac newMac() {
-        try {
-            SecretKeySpec keySpec = new SecretKeySpec(randomKey, ALGORITM);
-            Mac mac = Mac.getInstance(ALGORITM);
-            mac.init(keySpec);
-            return mac;
-        } catch (NoSuchAlgorithmException | InvalidKeyException exception) {
-            throw new IllegalStateException(exception);
-        }
-    }
-
     @Override
     public ByteBuffer newId(ByteBuffer buffer, int length) {
         ObjectUtil.checkNotNull(buffer, "buffer");
         ObjectUtil.checkPositive(buffer.remaining(), "buffer");
         ObjectUtil.checkInRange(length, 0, maxConnectionIdLength(), "length");
 
-        Mac mac = MACS.get();
-        mac.reset();
-        mac.update(buffer);
-        byte[] signBytes = mac.doFinal();
-        if (signBytes.length != length) {
-            signBytes = Arrays.copyOf(signBytes, length);
-        }
-        return ByteBuffer.wrap(signBytes);
+        return Hmac.sign(buffer, length);
     }
 
     @Override
