+++ codec-native-quic/pom.xml
@@ -56,7 +56,7 @@
     <quicheHomeIncludeDir>${quicheHomeDir}/quiche/include</quicheHomeIncludeDir>
     <quicheRepository>https://github.com/normanmaurer/quiche</quicheRepository>
     <quicheBranch>c_api</quicheBranch>
-    <quicheCommitSha>b140659a16bba84d2a30d8990ec5c9c9b0b88104</quicheCommitSha>
+    <quicheCommitSha>2893a3b3837ca11350fc7f542901c56a3ae49e2c</quicheCommitSha>
     <generatedSourcesDir>${project.build.directory}/generated-sources</generatedSourcesDir>
     <templateDir>${project.build.directory}/template</templateDir>
     <cargoTarget />
